+++ src/cli/com/yetanalytics/persephone/cli/match.clj
@@ -2,7 +2,8 @@
   (:require [com.yetanalytics.persephone :as per]
             [com.yetanalytics.persephone.cli.util.args :as a]
             [com.yetanalytics.persephone.cli.util.file :as f]
-            [com.yetanalytics.persephone.cli.util.spec :as s]))
+            [com.yetanalytics.persephone.cli.util.spec :as s])
+  (:import [clojure.lang ExceptionInfo]))
 
 (def match-statements-options
   [["-p" "--profile URI"
@@ -20,13 +21,15 @@
     :validate  [s/iri? s/iri-err-msg]
     :update-fn (fnil conj [])]
    ["-s" "--statement URI"
-    "Statement filepath/location; must specify one or more."
+    "Statement filepath/location; must specify one or more. Accepts arrays of Statements."
     :id        :statements
     :missing   "No Statements specified."
     :multi     true
     :parse-fn  f/read-statement
-    :validate  [s/statement? s/statement-err-msg]
-    :update-fn (fnil conj [])]
+    :validate  [s/statements? s/statements-err-msg]
+    :update-fn (fn [xs s]
+                 (let [xs (or xs [])]
+                   (if (vector? s) (into xs s) (conj xs s))))]
    ["-n" "--compile-nfa"
     (str "If set, compiles the Patterns into a non-deterministic finite "
          "automaton (NFA) instead of a deterministic one, allowing for "
@@ -39,17 +42,20 @@
    match failures or errors and return `false` if errors or failures exist,
    `true` otherwise."
   [{:keys [profiles pattern-ids statements compile-nfa]}]
-  (let [compiled (per/compile-profiles->fsms
-                  profiles
-                  :validate-profiles? false
-                  :compile-nfa?       compile-nfa
-                  :selected-patterns  (not-empty pattern-ids))
-        state-m  (per/match-statement-batch compiled
-                                            nil
-                                            statements
-                                            :print? true)]
-    (not (boolean (or (-> state-m :error)
-                      (-> state-m :rejects not-empty))))))
+  (try
+    (let [compiled (per/compile-profiles->fsms
+                    profiles
+                    :compile-nfa?       compile-nfa
+                    :selected-patterns  (not-empty pattern-ids))
+          state-m  (per/match-statement-batch compiled
+                                              nil
+                                              statements
+                                              :print? true)]
+      (not (boolean (or (-> state-m :error)
+                        (-> state-m :rejects not-empty)))))
+    (catch ExceptionInfo e
+      (s/handle-asserts e)
+      false)))
 
 (defn match
   "Perform Pattern matching based on `arglist`; print match failures or errors